# Food-Delivery-Microservice/Node-Microservices/ReviewService/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order API",
    "description": "API for managing Orders",
    "version": "1.0.0"
  },
  "host": "localhost:5300/api/review",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/menuId/{menuId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "menuId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/delete/{reviewId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}